fake news extension (2)/fake news extension/fake news extension/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
numpy==1.26.4
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.3
groq==0.0.13
flask_cors==3.0.10